requirements.txt
Django>=4.0
pandas
numpy<2.0
scikit-learn
scikit-surprise
requests
